break out ace.py from blackbox.py
